requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
reportlab>=4.2.0
pydantic>=2.10.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-auth
PyPDF2
